requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.18.0
requests>=2.31.0
sendgrid>=6.10.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
